+++ test/unit/sql/enum.test.js
@@ -5,7 +5,8 @@ var Support   = require(__dirname + '/../support')
   , DataTypes = require(__dirname + '/../../../lib/data-types')
   , expectsql = Support.expectsql
   , current   = Support.sequelize
-  , sql       = current.dialect.QueryGenerator;
+  , sql       = current.dialect.QueryGenerator
+  , expect    = require('chai').expect;
 
 
 describe(Support.getTestDialectTeaser('SQL'), function() {
@@ -24,6 +25,22 @@ describe(Support.getTestDialectTeaser('SQL'), function() {
         }
       });
 
+      describe('pgEnumName', function() {
+        it('does not add schema when options: { schema: false }', function() {
+          expect(sql.pgEnumName(PublicUser.getTableName(), 'mood', { schema: false }))
+            .to.equal('"enum_users_mood"');
+          expect(sql.pgEnumName(FooUser.getTableName(), 'theirMood', { schema: false }))
+            .to.equal('"enum_users_theirMood"');
+        });
+
+        it('properly quotes both the schema and the enum name', function() {
+          expect(sql.pgEnumName(PublicUser.getTableName(), 'mood', PublicUser.rawAttributes.mood.type))
+            .to.equal('"public"."enum_users_mood"');
+          expect(sql.pgEnumName(FooUser.getTableName(), 'theirMood', FooUser.rawAttributes.mood.type))
+            .to.equal('"foo"."enum_users_theirMood"');
+        });
+      });
+
       describe('pgEnum', function () {
         it('uses schema #3171', function () {
           expectsql(sql.pgEnum(FooUser.getTableName(), 'mood', FooUser.rawAttributes.mood.type), {
@@ -31,9 +48,9 @@ describe(Support.getTestDialectTeaser('SQL'), function() {
           });
         });
 
-        it('does not add schema when public', function () {
+        it('does add schema when public', function () {
           expectsql(sql.pgEnum(PublicUser.getTableName(), 'theirMood', PublicUser.rawAttributes.mood.type), {
-            postgres: 'CREATE TYPE "enum_users_theirMood" AS ENUM(\'happy\', \'sad\');'
+            postgres: 'CREATE TYPE "public"."enum_users_theirMood" AS ENUM(\'happy\', \'sad\');'
           });
         });
       });
@@ -45,8 +62,8 @@ describe(Support.getTestDialectTeaser('SQL'), function() {
           });
         });
 
-        it('uses the default schema if options given', function () {
-            expectsql(sql.pgListEnums(), {
+        it('uses the default schema if no options given', function () {
+          expectsql(sql.pgListEnums(), {
             postgres: 'SELECT t.typname enum_name, array_agg(e.enumlabel) enum_value FROM pg_type t JOIN pg_enum e ON t.oid = e.enumtypid JOIN pg_catalog.pg_namespace n ON n.oid = t.typnamespace WHERE n.nspname = \'public\' GROUP BY 1'
           });
         });